+++ acme2certifier/dbhandlers/wsgi_handler.py
@@ -445,7 +445,7 @@ class DBstore(object):
         """)
         self.logger.debug("create challenge")
         self.cursor.execute("""
-            CREATE TABLE "challenge" ("id" integer NOT NULL PRIMARY KEY AUTOINCREMENT, "name" varchar(15) NOT NULL UNIQUE, "token" varchar(64), "authorization_id" integer NOT NULL REFERENCES "authorization" ("id"), "expires" integer, "type" varchar(15) NOT NULL, "keyauthorization" varchar(128), "source" varchar(128), "status_id" integer NOT NULL REFERENCES "status" ("id"), "validated" integer DEFAULT 0, "validation_error" text, "created_at" TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL)
+            CREATE TABLE "challenge" ("id" integer NOT NULL PRIMARY KEY AUTOINCREMENT, "name" varchar(15) NOT NULL UNIQUE, "token" varchar(64), "authorization_id" integer NOT NULL REFERENCES "authorization" ("id"), "expires" integer, "type" varchar(15) NOT NULL, "keyauthorization" varchar(128), "source" varchar(128), "challenge_message_id" varchar(255), "status_id" integer NOT NULL REFERENCES "status" ("id"), "validated" integer DEFAULT 0, "validation_error" text, "created_at" TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL)
         """)
         self.logger.debug("create certificate")
         self.cursor.execute("""
@@ -605,6 +605,12 @@ class DBstore(object):
             self.logger.info("alter challenge table - add source‚")
             self.cursor.execute(
                 """ALTER TABLE challenge ADD COLUMN source varchar(128)"""
+            )
+
+        if "challenge_message_id" not in challenges_column_list:
+            self.logger.info("alter challenge table - add challenge_message_id")
+            self.cursor.execute(
+                """ALTER TABLE challenge ADD COLUMN challenge_message_id varchar(255)"""
             )
 
     def _db_update_cliaccount(self):
@@ -1593,9 +1599,12 @@ class DBstore(object):
         if "source" not in data_dic:
             data_dic["source"] = lookup["source"]
 
+        if "challenge_message_id" not in data_dic:
+            data_dic["challenge_message_id"] = lookup.get("challenge_message_id")
+
         self._db_open()
         self.cursor.execute(
-            """UPDATE challenge SET status_id = :status, keyauthorization = :keyauthorization, source= :source, validated = :validated WHERE name = :name""",
+            """UPDATE challenge SET status_id = :status, keyauthorization = :keyauthorization, source = :source, challenge_message_id = :challenge_message_id, validated = :validated WHERE name = :name""",
             data_dic,
         )
         self._db_close()