+++ acme2certifier/django_app/models.py
@@ -97,6 +97,7 @@ class Challenge(models.Model):
     created_at = models.DateTimeField(auto_now_add=True)
     keyauthorization = models.CharField(max_length=128, blank=True)
     source = models.CharField(max_length=128, blank=True)
+    challenge_message_id = models.CharField(max_length=255, blank=True)
     validated = models.IntegerField(default=0)
     validation_error = models.TextField(blank=True)